Add C ABI ffi between rust crypto lib and espressocrypto go module.
This commit adds FFI support between exsisting Rust crypto helper code
and the exsisting espressocrypto go module in native.go These functions
will now serve as utility functions to use in the integration that will
handle ffi to the Rust code.

This commit also moves the espresso-crypto-helper code to the
espressocrypto directory as we no longer need it for the arbitrator.