Allow 66 seconds for the startup of the inbox reader
In order to allow the inbox reader to read the init message start with a
polling time of roughly 1 minute against the parent-chain sequencer
inbox. This is because we need to populate the init message in the db
for the start method to work, but we won't do that if the initial
iteration of the run method takes longer than this timeout.