Add non-zero delay time for finality check (#655)
When the run loop that checks for finality returns a duration of zero to the underlying
goroutine run-loop, it effectively informs the runloop to reschedule immediately.  Since
the `pollSubmittedTransactionForFinality` will return `0` in all successful checks, and
since "no work to be done" is considered not to be an error, this means that this loop
will continuously re-evaluate.

The shared Espresso lock is acquired each time this is run and then released, making
this method the most likely for lock contention in terms of preventing others from
acquiring the lock.

There are more lock considerations to be performed, but this is a start for now.

(cherry picked from commit 49c4303789e61465b908fa2f5e78f403ee84e999)