Refactor test setup to allow for easier mock setup
The Test environment that is being setup to measure performance is quite flexible
but it is not obvious how to setup effectively.  It currently requires many individual
calls and configurations that allow for greater flexibility, but not the ease of use and
setup.  Ideally, we would like the ability to setup a mock environment with some
default configuration without the need to specify things.

In other words, the ideal environment should be able to be setup with the invocation of
a single function, while still allowing for specific configurations to be modified.

This change refactors the aspects of this setup in an effort to provide the caller with
maximum flexibility, and control.