Add UnimplementedEspressoClient
In order to add flexibility to our testing approach, we need to control certain elements
of our environment.  One of these elements is the EspressoClient itself.  The
EspressoClient serves as an interface that allows the consumer to interact with the
Espresso Chain.  This is usually done via an API request, but it doesn't necessarily
need to mean that.

Certain aspects of the EspressoClient's behavior could potentially have a profound
impact our implementation and as such we need to be able to control some aspects
about the EspressoClient, without necessarily hitting a live Espresso Chain,  or
worrying about all of the other aspects of the EspressoClient.

As a result, and to make it easier to implement some future implementations the
UnimplementedEspressoClient has been defined in order to make it easier to
write Mocks of the EspressoClient without the need to fulfill the full interface.

(cherry picked from commit f6949b2cf5cd6391ac71aacf02cfa6319733aa15)