Add UnimplementedLightClientReader
In order to add flexibility to our testing approach, we need to control certain elements
of our environment.  One of these elements is the LightClientReaderInterface.  The
LightClientReaderInterface represents a type that is interacting with a smart contract
on an RPC endpoint, and reading it's relevant state.

Certain aspects of the LightClientReaderInterface will affect how our logic progresses
and it is beneficial to be able to have direct control over the results this interface
returns.

As a result, and to make it easier to implement some future implementations the
UnimplementedLightClientReader has been defined in order to make it easier to
write Mocks of the LightClientReaderInterface without the need to fulfill the full
interface.

(cherry picked from commit e75fdd7a79882380435bd606d8408144fb064b25)