Add MockExecutionEngineForTransactionStreamer
The `TransactionStreamer` utilizes an `ExecutionClient` in its operation, but it only
calls a subset of the `ExecutionClient` interface.  While the `ExecutionClient` plays
an important role, we can provide a simplified Mock implementation that only needs to
cater to the specifics, as understood by the `TransactionStreamer`.

The `MockExecutionEngineForTransactionStreamer` does just this by implementing
the minimum methods needed for the `TransactionStreamer` in an effort to provide
the implementation with the simplest implementation possible.  This **should** help
to eliminate any `ExecutionClient` implementation errors / details from consideration
with the Espresso specific interaction of the `TransactionStreamer`.

(cherry picked from commit 56fdcf26f3f7ad0a409f868a7fbbf9441d7e9e51)