Add MockAlwaysLiveLightClientReader
The `TransactionStreamer` currently only ever utilizes the
`LightClientReaderInterface` in order to check liveness of `HotShot`.  In the majority
of our test cases, we want `HotShot` to be considered live. As a result, it is beneficial
to remove the complications of setting up Smart Contract reading that is dependent
on network and chain conditions, and to instead just utilize a replacement that just
always considers `HotShot` as live, with no delay in communication.

(cherry picked from commit 02702f5e5ca621f11009d47059b7c5077d4a33a7)