Add SiphonBlocksWithTransactions
When interacting with the Espresso Chain, it becomes beneficial to be able to inspect
what is being retrieved from the Espresso Chain itself.  This allows one the ability to
inspect the response, and check it against the logic that consumes in other aspects
such as the resubmission, and finality checking processes.

In order to facilitate such a need, the `SiphonBlocksWithTransactions` has been
written to provide this data so that the individual `TransactionsInBlock` responses
can be inspected, and checked for consistency.

This also has the added benefit of knowing immediately when this data is being
provided, and should be considered "available" for the system as a whole.

(cherry picked from commit 833452a729f98c7cfb5a94c4d19a0c3097d3de2f)