Add MessageGenerator
In order to be able to test the system effectively in a simulated scenario, we need to
be able to easily submit messages into the system so that they can simulate load
in an easily controllable manner.

We add the ability to easily generate new messages sequentially to test against the
`TransactionStreamer`.  These messages, can be generated and submitted at
controlled rates that can be further expanded in the future.

(cherry picked from commit 5806afa7ac601f23ec51428b834b1b4474abc960)