Add TimingData nd Timing function / constructor
In order to effectively time things, it helps to have a convenient and consistent way to
represent timing data. The timing data is meant to be as simple as possible, and is not
meant to be constructed directly, though there is nothing stopping the user from doing
just that.

(cherry picked from commit 3eb29ac84b89c743cad01926add3be5937c1cf91)