Split Espresso code from transaction_streamer
In order to make the management of the Espresso integration from the rest of the
code, it is helpful to have the Espresso additions as separate and isolated from the
rest of the code as possible.  In this way all of the Espresso changes becomes as
additive as possible.  This can help to make the code being added easier to review
in isolation, only needing to inspect the bridge boundaries at the point where they
interact.

This change makes an effort to split out the Espresso submission process from the
rest of the existing code base into its own self-contained areas.  In doing this, the
code should be better organized, while providing an interface for the boundaries
where these two separate systems communicate.

NOTE: This can be further refined, especially around the area of configuring and
registering the TEE

(cherry picked from commit 280d41a9d962a67ddd9f6d61f7fbae8b8eb81335)