Add BlockBuilder for MockEspressoChain
In order to simulate real-world conditions it is beneficial to be able to simulate a
non-ideal block production that we currently see on the Espresso Chain.  The Blocks
that Espresso produces are essentially size constrained in order to ensure that the
production process doesn't tax the system more than it needs to be.

These restrictions can have an impact on the throughput rates we are able to
achieve.

As a result, it is beneficial to be able to simulate these restrictions in order to more
closely compare oujr tests against real-world scenarios.