Revise TestTransactionStreamerEspressoThroughput conditions
The TestTransactionStreamerEspressoThroughput test originally was created just to
monitor throughput based on the parameters being configured for the mock
setups.  The idea was to determine how interval configuration impacted the integration
performance under idealized conditions.

With some base lines set, it is now important to try and ensure that our integration
continues to perform at an established baseline and well beyond in order to ensure
that we ware able to catch up in a sufficient amount of time, should we fall behind.

To achieve this goal the TestTransactionStreamerEspressoThroughput test has been
reworked to test specific scenarios against the required benchmarks.  Additionally
it is important to codify the requirements of it being able to outperform the baseline
by a factor.  In this case the factor chosen is around 8x, but this was chosen arbitrarily
and our tests have indicated that we are able to exceed this with the right polling
interval configuration.