Fix error cases where submitted transaction tracking is lost
There are a few edge cases that have been identified that would result in explicit
submitted transaction tracking loss:
- If a transaction is found successfully mapped to
a block height, but then the retrieval of the transactions for that block fail, we lose track of the transaction
- If a transaction resubmit attempt fails, then we lose track of the transaction

To handle these cases we should fallback to continue tracking the existing submission
we already have.  This will ensure, at the very least, that we don't ever drop
transactions that we are actively still trying to verify before we are meant to when
we encounter this edge cases.