Rename Throughput test
In order to ensure that the Throughput is picked up by ci, it's beneficial to have its
name follow the existing `run-espresso-ci-job` scheme.  That is that it's name
should contain `TestEspresso`.  This alone won't ensure that it's evaluated in ci, but
it helps to be consistent in order to identify these tests more effectively.