Modify tests being targeted by run-espresso-ci-job
The `run-espresso-ci-job` will only target tests that contain the name `TestEspresso`
in them, and are contained directly within the `system_tests` directory.  It would be
beneficial to target any `TestEspresso` tests that are a descendent of the `system_tests`
instead of just residing directly within it.

This change expands the tests being considered to be any descendents of `system_tests`
rather than just sitting in that single folder.