Remove `sudo` from `prepare-espresso-crypto-helper` script
The script ./scripts/prepare-espresso-crypto-helper utilizes a remotely executed
script to in order to help download and make available a binary library that is needed
for this project.  This script has been modified in such a way to no longer require
privilege escalation in order to operate properly. As a result, the
`prepare-espresso-crypto-helper` script needs to be updated to utilize the new
command, which can be installed locally.

This modifies the script to utilize the new command, and provide instructions to
the user, should the command not exist locally.

NOTE: This will **NOT** install the command for the user automatically.  Instead it
provides instructions to the user to ensure that it is installed locally, so that the user
can evaluate the trust being put into the external utility, and will ensure that it
doesn't get updated or replaced without his/her direct interaction.