Separate the Nitro Message handling from Espresso Transactions
In order to make the MultiWorkerQueueEspressoSubmitter more agnostic and to not
have to be associated with the details of the Nitro network, it is beneficial to have
a clear boundary of separation where possible.

To facilitate this, the implementation has been separated into two different managing
structures, one that manages the Espresso Transaction Submission process in
isolation, and one that handles the Nitro Message submission, bundling, and
transaction submission scheduling.