Remove magic values in favor of configurable ones
There are a few values in the Multi Worker Queue solution that are set to hard coded
magic values.  In order to favor flexibility and configurability, these values should be
adjustable in some way.

This change adds configuration values and functional options for all of these values
and moves the current default values to the configuration defaults variable.