Relocate helpful functions to util
Addresses feedback from @Sneh1999:
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279764428
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279787412

There are a few helper functions utilities and types that have been introduced for the
multi worker queue espresso submitter that do not need to apply to it exclusively.
These helper functions have been suggested to move to a utilities location for ease of
use.

This change relocates these functions to the `util` package so they can be freely
utilized where needed.