Modify delay on job processing failure reattempts to be configurable
The goroutine workers that are dedicated to processing a single job / task at a time
have consideration for job failure with a sleep to slowly back off requests.  This delay
is currently hard-coded to a magic number of `100ms`.  This value should really be
configurable.

This change replaces these delay values with a value that is ultimately derived from
configuration.  This allows the values to be overwritten while still allowing for a
default value.