Modify Adapter to have positions be configurable
The `NitroMessageToEspressoTransactionAdapter` currently starts in a state where it
is unable to have any pre-existing knowledge of the data that is available or has been
submitted to Espresso.  As a result, it will always attempt to sync the messages from
position `0`.

The idea is to allow for the user to be able to specify the position that we'd like to
start submitting to Espresso.  Additionally, we would like for the user to be able to
indicate which message indexes are available.

This change allows the user to specify both of these values.