Update naming across project and get caff node test passing
This commit fixes the caff node test (from the POV of the caff node
itself) and updates naming conventions across the project to reflect the
new registerService paradigm.

Unforutnaltey, during the caff node test, the batch poster seems to
encounter execution reverts for transactions that attempt to post a
batch. I haven't figured out why this is quite yet today (I've been
juggling multiple items of work today) but I have an idea of what it
might be so hopefully we can fix this in a future PR.