Fix lint errors and attempt to fix caff node restart test
This commit fixes lint errors not previously illustrated in CI or in
local linter runs.
This commit also attempts to fix the caff node restart test. However, I
think we should take a closer look at what this test is really providing
us in terms of code coverage vs. the amount of time that it takes up in
our CI runners. It seems to me as if it is mainly testing the TEE
configurations in relation to the `NewEspressoCaffNode` method. If this
is the case, we can likely make a much more efficient test.