Register caff node auth db (#822)
* Register Caff node key in EspressoTEEVerifier contract

* get the caff node tests working

* lint and fix docker build

* remove nsm

* fix test

* fix tests

* add private key to env for hyperlane validator

* update espresso tee contracts submodule

* Rebase Sneh's work from #770 on release-tee-caff-node

* Allow caff node to run without registering when not configured to use a TEE

* Update `Makefile` for utilizing new binding generation in the espresso-tee-contracts submodule

This commit allows users build the espresso-tee-contracts specific
bindigns by running `make contracts`, and clean the artifacts via `make
clean`

* update go.mod and espresso-tee-contracts repo

* Attempt to use new bindings in batch poster

* Update `Makefile` to reflect new project structure in `espresso-tee-contracts`

* update makefile to properly handle bindings generated

* update go.mod file to remove unnecessary replace directive

* Update binding references in relevant locations.

* Fix file path in Makefile

* Update code in espressotee/ to interact with new contract bindings.

* Update

* Update  with new bindings.

* fix dependencies for integration

* fix test compilation

* custom patch with some commits to update caff node

* Update naming across project and get caff node test passing

This commit fixes the caff node test (from the POV of the caff node
itself) and updates naming conventions across the project to reflect the
new registerService paradigm.

Unforutnaltey, during the caff node test, the batch poster seems to
encounter execution reverts for transactions that attempt to post a
batch. I haven't figured out why this is quite yet today (I've been
juggling multiple items of work today) but I have an idea of what it
might be so hopefully we can fix this in a future PR.

* Update contract related submodules

* Update bold submodule

This commit updates the bold submodule to utilize the new
espresso-tee-contracts version, and a compatible `SequencerInbox`
version to allow the batch posting tests to work if we ever need to
implement bold.
For now, our integration does not, but it makes some sense to get ahead
of this.

* Fix lint issues and config creation

This commit fixes the duplicaton of the espresso-tee-type config option
in the espresso caff node, and updates it's description.
Additionally this fixes outstanding lint errors not previously caught by
the pre-commit hook, likely due to applying previous commits made with
`git commit -n` via cherry-picking or custom patches.

* Fix lint errors and attempt to fix caff node restart test

This commit fixes lint errors not previously illustrated in CI or in
local linter runs.
This commit also attempts to fix the caff node restart test. However, I
think we should take a closer look at what this test is really providing
us in terms of code coverage vs. the amount of time that it takes up in
our CI runners. It seems to me as if it is mainly testing the TEE
configurations in relation to the `NewEspressoCaffNode` method. If this
is the case, we can likely make a much more efficient test.

* Fix `TestEspressoCaffNodeRestartWithTeeType`

This commit fixes a caff node test that was intended to have a AuthDB
initialized with an HMAC value from the start. A previous commit
inadvertently changed this to have an AuthDB that didn't perform HMAC
verificaiton on it's first run, which caused it to fail on restart.

* Fix nit of extra newline in Dockerfile

* Update private key export env var

* Fix service type check in key manager

* update comment related to hardcoded service type in call to `Verify()`

* Add tee verifier address to config for cross chain NFT project

* add description to tee verifier config option

* Add tee type check in nitro.go

* fix test service type value in types.go

* update naming of config EspressoTEEVerifier contract

---------

Co-authored-by: Sneh Koul <snehkoul1999@gmail.com>