Add implementation of multi queue worker proposal (#700) (#843)
* Revise TestTransactionStreamerEspressoThroughput conditions

The TestTransactionStreamerEspressoThroughput test originally was created just to
monitor throughput based on the parameters being configured for the mock
setups.  The idea was to determine how interval configuration impacted the integration
performance under idealized conditions.

With some base lines set, it is now important to try and ensure that our integration
continues to perform at an established baseline and well beyond in order to ensure
that we ware able to catch up in a sufficient amount of time, should we fall behind.

To achieve this goal the TestTransactionStreamerEspressoThroughput test has been
reworked to test specific scenarios against the required benchmarks.  Additionally
it is important to codify the requirements of it being able to outperform the baseline
by a factor.  In this case the factor chosen is around 8x, but this was chosen arbitrarily
and our tests have indicated that we are able to exceed this with the right polling
interval configuration.

* Add Multi Worker Queue Espresso Submitter

Based on potential issues identified in the polling espresso submitter, an alternative
approach has been proposed.  This approach focuses on performing different
parts of submitting transactions to espresso in discrete units of work that can be
scaled independently.  This would allow for multiple Espresso Transactions to be
inflight in terms of submission and verification at the same time, in order to maximize
the amount of throughput that is allowed.

To realize this approach this adds the `MultiWorkerQueueEspressoSubmitter` which
manages and runs two separate work queues for submitting and verifying transactions
that are submitted to Espresso.  The idea here is that this will only focus on the parts
related to Espresso as much as possible without needing to worry about the parts
that are specific to the underlying integration.

By taking this approach, the result should be a more reliable and trackable process for
submitting and verifying transactions to the Espresso Network.

NOTE: There is still work to be done here with regards to adapting the specific
integration to the transactions being submitted to Espresso, and retrieving them
from Espresso.

This also updates the `TestTransactionStreamerEspressoThroughput` test to measure
the throughput performance with targets as compared to the Polling Espresso
Submitter solution.

* Separate the Nitro Message handling from Espresso Transactions

In order to make the MultiWorkerQueueEspressoSubmitter more agnostic and to not
have to be associated with the details of the Nitro network, it is beneficial to have
a clear boundary of separation where possible.

To facilitate this, the implementation has been separated into two different managing
structures, one that manages the Espresso Transaction Submission process in
isolation, and one that handles the Nitro Message submission, bundling, and
transaction submission scheduling.

* Remove magic values in favor of configurable ones

There are a few values in the Multi Worker Queue solution that are set to hard coded
magic values.  In order to favor flexibility and configurability, these values should be
adjustable in some way.

This change adds configuration values and functional options for all of these values
and moves the current default values to the configuration defaults variable.

* Relocate helpful functions to util

Addresses feedback from @Sneh1999:
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279764428
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279787412

There are a few helper functions utilities and types that have been introduced for the
multi worker queue espresso submitter that do not need to apply to it exclusively.
These helper functions have been suggested to move to a utilities location for ease of
use.

This change relocates these functions to the `util` package so they can be freely
utilized where needed.

* Split file into multiple files

Based on feedback from @Sneh1999 and @zacshowa:
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279797524
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#pullrequestreview-3084968050

With the added functionality and types introduced for multi worker queue processing
the solution has become somewhat large and can make it more difficult to navigate
the implementation.  The suggestion has been put forth to separate the large file
into separate files.

This change separates each of the primary types into their own files for isolated
references where possible.  The errors have been kept next to their place of utilization
for faster lookup.

* Update forloop to make the endPos inclusive

Based on feedback provided by @Sneh1999:
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279769374

There is an end case in the bundleTransactions function where payloads may not be
processed as expected.  The specific case is where `startPos` == `endPos`.

This change fixes the for loop to include the `endPos`.

* Modify log levels

Based on feedback from @Sneh1999 and @zacshowa:
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279791212
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279791489
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279794032
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279795365
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279795586
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279796342
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279782182
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2279782298

In order to identify potential issues as they occur in production it is beneficial to be able to identify when certain events that hinder the ability for our solution to operate.
Based on this several log levels have been changed based on requests, and others
have been added.

* Fix typo

Based on feedback from @zacshowa:
https://github.com/EspressoSystems/nitro-espresso-integration/pull/700#discussion_r2252008810

The word "utilities" is used instead of "utilizes".

This change corrects the wrong word being used.

* Modify delay on job processing failure reattempts to be configurable

The goroutine workers that are dedicated to processing a single job / task at a time
have consideration for job failure with a sleep to slowly back off requests.  This delay
is currently hard-coded to a magic number of `100ms`.  This value should really be
configurable.

This change replaces these delay values with a value that is ultimately derived from
configuration.  This allows the values to be overwritten while still allowing for a
default value.

* Modify Adapter to have positions be configurable

The `NitroMessageToEspressoTransactionAdapter` currently starts in a state where it
is unable to have any pre-existing knowledge of the data that is available or has been
submitted to Espresso.  As a result, it will always attempt to sync the messages from
position `0`.

The idea is to allow for the user to be able to specify the position that we'd like to
start submitting to Espresso.  Additionally, we would like for the user to be able to
indicate which message indexes are available.

This change allows the user to specify both of these values.

* Fix lint issues

---------


(cherry picked from commit 48c7252ba2f633eb27b424a80345917f3ea51149)

Co-authored-by: Theodore Schnepper <Ayiga@MSN.com>